requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
distro==1.9.0
Django==5.2.6
dnspython==2.8.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
image==1.5.33
jiter==0.10.0
openai==1.107.1
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.15.0
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.47.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
